demo: move all *_demo.py scripts from vultron/scripts/ to vultron/demo/ (DEMO-4.3)
Move 12 demo scripts to vultron/demo/ so they are importable as
vultron.demo.<script_name> per DC-02-002.

- Update test/scripts/ imports: vultron.scripts.X_demo → vultron.demo.X_demo
- Update docker/Dockerfile CMDs to use vultron.demo module paths
- All demos retain if __name__ == '__main__' invocability

568 tests pass.
