+++ test/scripts/test_manage_case_demo.py
@@ -17,7 +17,7 @@ from _pytest.monkeypatch import MonkeyPatch
 from fastapi.testclient import TestClient
 
 from vultron.api.main import app as api_app
-from vultron.scripts import manage_case_demo as demo
+from vultron.demo import manage_case_demo as demo
 
 
 @pytest.fixture(scope="module")
